Rework most methods, adding the global scaffolding for testing
